fix: test폴더 이름 CarsTest -> CarTest 변경
